Email notification working

+++ BatchDataProcessing/src/main/java/com/training/bean/FailedProduct.java
@@ -10,13 +10,13 @@ public class FailedProduct implements Serializable{
 	private String orgId;
 	private String configId;
 	private String releaseDate;
-	private long messageId;
+	private int messageId;
 	
 	public FailedProduct() {
 		super();
 	}
 	public FailedProduct(String upc, String productDesc, String artistId, String orgId, String configId,
-			String releaseDate, long messageId) {
+			String releaseDate, int messageId) {
 		super();
 		this.upc = upc;
 		this.productDesc = productDesc;
@@ -27,7 +27,7 @@ public class FailedProduct implements Serializable{
 		this.messageId = messageId;
 	}
 	// another constructor for setting product object to failedProduct parameters
-	public FailedProduct(Product product, long messageId) {
+	public FailedProduct(Product product, int messageId) {
 		super();
 		this.upc = product.getUpc();
 		this.productDesc = product.getProductDesc();
@@ -73,10 +73,10 @@ public class FailedProduct implements Serializable{
 	public void setReleaseDate(String releaseDate) {
 		this.releaseDate = releaseDate;
 	}
-	public long getMessageId() {
+	public int getMessageId() {
 		return messageId;
 	}
-	public void setMessageId(long messageId) {
+	public void setMessageId(int messageId) {
 		this.messageId = messageId;
 	}
